+++ deploy.sh
@@ -1,4 +1,4 @@
-for VAR in PERSONAL_API_KEY MANIFEST_FILE_eval APP_ID_eval APP_NAME_eval  MANIFEST_FILE_api_periodic APP_ID_api_periodic_acc APP_ID_api_periodic_prod APP_NAME_api_periodic_acc APP_NAME_api_periodic_prod MANIFEST_FILE_api_on_demand APP_ID_api_on_demand_acc APP_ID_api_on_demand_prod APP_NAME_api_on_demand_acc APP_NAME_api_on_demand_prod MANIFEST_FILE_admin APP_ID_admin APP_NAME_admin;
+for VAR in PERSONAL_API_KEY MANIFEST_FILE_eval APP_ID_eval APP_NAME_eval  MANIFEST_FILE_api_periodic APP_ID_api_periodic_acc APP_ID_api_periodic_prod APP_NAME_api_periodic_acc APP_NAME_api_periodic_prod MANIFEST_FILE_api_on_demand APP_ID_api_on_demand_acc APP_ID_api_on_demand_prod APP_NAME_api_on_demand_acc APP_NAME_api_on_demand_prod MANIFEST_FILE_admin APP_ID_admin APP_NAME_admin MANIFEST_FILE_discharge APP_ID_discharge APP_NAME_discharge;
 do
     if [[ -z ${!VAR+x} ]]; then
         echo "env variable" $VAR "missing"
@@ -6,7 +6,7 @@ do
     fi
 done
 
-read -p "What do you want to deploy? Options: 'eval'/1 ; 'api-periodic-acc'/2 ; 'api-periodic-prod'/3  ; 'api-on-demand-acc'/4 ; 'api-on-demand-prod'/5 ; 'admin'/6 - " APPLICATION
+read -p "What do you want to deploy? Options: 'eval'/1 ; 'api-periodic-acc'/2 ; 'api-periodic-prod'/3  ; 'api-on-demand-acc'/4 ; 'api-on-demand-prod'/5 ; 'admin'/6 ; 'discharge-dashboard'/7 ;" APPLICATION
 
 read -p "Have you updated the environment variables in Posit Connect? (y/n)" ANSWER
 if [[ $ANSWER == 'n' || $ANSWER == "N" ]]; then
@@ -46,6 +46,11 @@ elif [[ $APPLICATION == 'admin' || $APPLICATION == "6" ]]; then
     cp $MANIFEST_FILE_admin manifest.json
     APP_ID=$APP_ID_admin
     APP_NAME=$APP_NAME_admin
+elif [[ $APPLICATION == 'discharge-dashboard' || $APPLICATION == "7" ]]; then
+    echo "Deploying discharge dashboard"
+    cp $MANIFEST_FILE_discharge manifest.json
+    APP_ID=$APP_ID_discharge
+    APP_NAME=$APP_NAME_discharge
 else
     echo "Invalid option"
     return