+++ src/discharge_docs/api/api_helper.py
@@ -30,6 +30,7 @@ class ApiEndpoint(Enum):
 def check_authorisation(key: str, stored_key: str) -> None:
     # Check if the provided key matches the stored key in the environment variables.
     if key != os.environ[stored_key]:
+        logger.error("403: Invalid API key")
         raise HTTPException(
             status_code=403, detail="You are not authorized to access this endpoint"
         )