+++ src/discharge_docs/config.py
@@ -1,8 +1,11 @@
+import logging
 import os
+import sys
 from pathlib import Path
 
 import tomli
 from pydantic import BaseModel
+from rich.logging import RichHandler
 
 CONFIG_PATH = Path(__file__).parent / "llm" / "deployment_config.toml"
 
@@ -22,6 +25,34 @@ def load_config(config_path: Path = CONFIG_PATH) -> LLMConfig:
         f"deployment_name_{os.getenv('ENVIRONMENT', 'acc')}"
     ]
     return LLMConfig(**deployment_config_dict, deployment_name_env=deployment_name_env)
+
+
+def setup_root_logger() -> None:
+    """Setup the root logger for the project.
+    This function sets up the root logger with a specific format and level.
+    """
+    root_logger = logging.getLogger()
+    root_logger.setLevel(logging.INFO)
+    # Some packages have already initialized the root logger, so we need to remove
+    # their handlers first.
+    for handler in root_logger.handlers[:]:
+        root_logger.removeHandler(handler)
+
+    # Posit does not support rich logging, so use a simple console handler
+    if os.getenv("CONNECT_SERVER") is not None:
+        console_handler = logging.StreamHandler(sys.stdout)
+        console_handler.setFormatter(
+            logging.Formatter("%(levelname)s: [%(name)s] - %(message)s")
+        )
+    else:
+        console_handler = RichHandler()
+        console_handler.setFormatter(
+            logging.Formatter(
+                "%(message)s",
+                datefmt="[%X]",
+            )
+        )
+    root_logger.addHandler(console_handler)
 
 
 config = load_config(CONFIG_PATH)